Refactor how download classes are structured
The soon-to-come ClickNDownload downloader needs finer control over both
the initial request and the streaming downnload request to add additional
args that are awkward to handle with the structure that exists.

Instead, the terminal download classes will be responsible for implementing
task_impl().  FileDownloader will implement a helper to handle taking a
streaming response and saving it to a file.  That means the termina classes
will have to handle creating and cleaning up that streaming request.

Moved the timeout/retry wrapper for http requests into the Task class, since
that's where the client is created.

The Kraken and GoFile don't have to change all that much to support this
change.