Remove the top-level references to the work queue
In order to make it possible for a concrete downloader to throw an exception
signalling it's intent to give up and that the system should try another
supported downloader, we're going to be refactoring the whole thing to use
a semaphore to limit the number of simultaneous downloads, rather than the
workers+queue currently in use.

In the current design, once a job is placed in the queue and started by a
worker, there's no reasonable way for the original download-picker to catch
that exception and try another.