A big refactor to get things working without the task queue
There are a ton of changes and file moves, and yes it's hard to follow from
just looking at the diffs.  I actually copied all the files here to another
repo and hacked at it until things started working.  This commit represents
the point where it all started working again:
    * Search for something and we could resolve that to a Tosho ID
    * Look up that ID and get a list of sources+links
    * Download something successfully from KrakenFiles

Code common to httpx AsyncClients is moved to HttpClient.py

What used to be in PickFileDownloadSource.py (a task to pick a source to
download a file from and queue that download from a specific downloader)
now lives in FileDownloader.py and represents the intent to download a whole
file, even if it's split into pieces and available from many sources.

FileDownloader used to be the base class for source-specific downloaders
is now moved to DownloadSourceBase