Properly throttle the number of downloads with the semaphore
I wasn't using the semaphore properly before.  The only code path protected
by the semaphore was creation of the download task, and that lead to all the
downloads of a batch starting concurrently, with many of them dying of
various timeouts.

The right way is to have the semaphore protect the actual download which is
blocked with an await