Try another source if there's problems connecting
After exhausing the retries connecting/downloading from one source, try the
next one instead of killing the whole program.

Part of this is a bugfix where download() should have been async from the
start, necessitating calling await on the exception_retry().  This bug was a
pain to track down because the non-async download() meant that the call to
dl.download() in FileDownloader.download_piece() actually vanishes from the
call stack and appears as
    FileDownloader.download_piece()
    calls exception_retry()
    calls ActualDownloadClass.download_from_url()
and there's no opportunity to catch the exception within
DownloadSourceBase.download()